Platform/ARM: Match asl compiler output for Juno
The iASL compiler's -tc option has been updated to generate
output that is suitable for including from a C file. The iASL
compiler uniquely names the C array containing the processed
table data as <ASL filename>_aml_code.The patch for this update,
'iASL: Enhance the -tc option (create AML hex file in C)',
dated 16 March 2018 (2018-03-16), can be seen in the ACPICA git
repository at
https://github.com/acpica/acpica/commit/f9a88a4c1cd020b6a5475d63b29626852a0b5f37

A corresponding change is required in the ConfigurationManager
and this patch contains the necessary changes to adapt to the
generated iASL compiler output.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>