Platform/ARM: Add OEM CPU generator for FVP
Add support for dynamic generation of ACPI CPU device information.
This generator uses the compiled data from a template asl file and
patches it at runtime to generate the CPU information based on the
number of CPUs and their ACPI UID. This patched data is then
installed as a SSDT table.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>