Platform/ARM: Juno: Config Mgr remove GICD ID
According to ACPI 6.2 Specification - Errata A, 'One,
and only one, GIC distributor structure must be present
in the MADT for an ARM based system'. Accordingly the
CM_ARM_GICD_INFO structure has been updated to remove
the GicId field as this value is set to zero by the
MADT generator.

This update reflects the corresponding change required
to the Configuration Manager.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>