Platform/ARM: FVP: Enable SSDT Serial generation
The SSDT Serial Port Generator queries the
configuration manager for a list of serial
port devices. It uses the AmlLib library
interfaces to load a Serial Port AML template
code and represents it as an AML tree.
It then traverses the AML tree to locate and
fixup the Base address and Interrupt information
for the serial port devices. The AML tree is
then serialized to a buffer representing the
definition block which is installed as an ACPI
SSDT table.

The SSDT Serial Port Generator queries the
configuration manager for a list of serial
port devices.

This patch enables dynamic SSDT Serial Port
generation and adds support to the configuration
manager to return the CM_ARM_SERIAL_PORT_INFO
objects that represents the serial port devices
on the platform.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>
Signed-off-by: Pierre Gondois <pierre.gondois@arm.com>