Platform/ARM/VExpressPkg: Merge ASL and CfgMngr directories
Due to the following patch available in edk2:
  0a4aa20e8d446c2f5dd54f3a0a7ec4d52f0ebdb6
  BaseTools: Compile AML bytecode arrays into .obj file
a C array containing the AML bytecode is generated from
ASL files, and a .obj file containing this array is
generated.

The Configuration Manager is currently providing an extra
flag to the ASL compiler to generate a C array containing
AML bytecode. This C array is contained in a .hex file.
This .hex file is included where necessary.

This mechanism was necessary due to the impossibility
to describe dependencies between ASL files and C files.
This mechanism is not required anymore. Thus the ASL and
CfgMngr directories can me merged in one module.

This patch merge the two directories for the FVP.

Signed-off-by: Pierre Gondois <pierre.gondois@arm.com>
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>