Platform/ARM/VExpressPkg: FVP RevC SysID.Rev defintion
The Base Platform RevC model is a configuration of the Base Platform
that includes a PCIe subsystem. It also includes system IP like SMMUv3.

The Base Platform FVP model and the Base Platform RevC FVP model
can be differentiated by examining the System ID register. For the
FVP RevC model the SysID.Rev is 0x2. Therefore, add this definition
so that software can read the System ID register to detect a FVP
RevC model.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>