Platform/ARM/VExpressPkg: Memory map for FVP RevC model
FVP Base Platform RevC is a configuration of the Base Platform that
includes a SMMUv3 and a PCIe subsystem. The PCIe subsystem has a
PCIe config region (ECAM) and two PCIe memory regions.

Add the SMMUv3 and PCIe config and memory regions to the memory map
so that the firmware can access and configure these components.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>