Platform/ARM/VExpressPkg: Add Virtio Block Device description
The FVP model implements a Virtio block device that can be accessed from
a Guest OS if it has an appropriate driver. It is targeted primarily at
Linux, which has a built-in Virtio block driver. VirtioBlockDevice allows
to use a file on the host that is specified using the models image_path
parameter, as a hard drive in the Guest OS.

Therefore, add the Virtio block device description to the DSDT.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>