Platform/ARM/VExpressPkg: ACPI support for FVP RevC model
Base Platform RevC is a configuration of the Base Platform that includes
a PCIe subsystem and a SMMUv3. It also has an AHCI-SATA disk controller
as a device on the PCIe bus.

To enable an OS to use these features add the following ACPI tables:
  - IORT
  - MCFG
  - SSDT PCIe

Also add checks such that these additional tables are only installed if
the FVP model is RevC. This allows a unified firmware for use with both
the FVP_Base_AEMv8A-AEMv8A model and the FVP_Base_RevC-2xAEMv8A model.

Note: The CPU affinities are shifted in the FVP_Base_RevC-2xAEMv8A model
and this is adjusted in InitializePlatformRepository().

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>